Always prefer user defined arbitraries when shrinking
If I have a test that assumes my arbitrarly generated set has length 1
(or some other property) and my test fails, the shrunken object needs to
have to have the same properties as the original object to be useful for
debugging. Shrinking the set to length 0 (or violating any expected
property) will cause a failed test but for a different reason than the
original failure.